neonUtil: Handle Instructors & Volunteers with no membership records
This is an unusual corner case but one easy to fix: existing code did not
fetch Neon accounts that were interesting due to the presence of "Instructor" or
"Volunteer" tags but didn't have any associated membership records.  With
this fix, new Instructors or Volunteers that weren't prior members will have
Openpath credentials issued with the appropriate permissions.