Error handling for blank teacher
When a teacher isn't assigned to a class yet, the "Event Topic" in Neon
will be blank. Because the result is an empty string, not `None`, the
script doesn't catch the error, and just silently fails. This update
catches any falsy value for the event topic, and should send an email to
classes@asmbly.org.

As it currently stands, the subject line will be "Failed Class Reminder
- ". See line 224 of the script - `TEACHER_EMAILS['']` will raise a
KeyError. We could fix this by changing line 224 to
`TEACHER_EMAILS.get(teacher)` (which would also make the overall
script a bit safer), but I'm not sure what the desired behavior is
yet.