Use split and truthy evaluation to avoid errors parsing teacher name
`.index(" ")` will throw an error if the target string isn't found, so
if teacher is `None` or `''`, this line will cause the whole script to
error out.

Instead, we can use `.split(' ')` and `if teacher`. `if teacher else
'N/A' accounts for the case where `teacher` is `None` (the result will
be the string `"N/A"`), and `.split(' ')` will handle any string cases,
including the empty string or a string where the given substring isn't
included.