Log exceptions, but send out remaining emails
This diffs adds two things: first, it wraps the body of the TEACHERS loop in a
try-except statement, which will allow us to log any exceptions that
happen while still sending out the remaining emails.

Secondly, it also wraps the majority of the script in a `main` function,
following best practices for Python scripts. This isn't strictly
necessary (and I can remove it if there's opposition), but it
does allow us to start breaking up the body of the script so that it's
easier to read and modify in the future.