Create framework for automated tests with GitHub Actions and Pytest
This doesn't actually add any tests yet, but gives us a relatively easy
way to start building out our testing infrastructure.

This also doesn't require tests to run before deploying main - I'd like
to add that in the future, but I don't want to risk interfering with the
deploy process until we're confident in the stability of our tests.