Add tests for dailyClassReminder.py (#39)
* Remove top-level API call to allow for mock responses

Top-level code is executed as soon as a module is imported, meaning that
it can't be mocked using our test framework. In order to write tests for
this code, we need to encapsulate the API call inside of a function so
that we can mock the API call rather than having it reach out to the
real API.

* Add passing tests for dailyClassReminders

This PR adds a MockEventBuilder and a few tests for the
dailyClassReminder script. All of these tests currently pass, so they
don't catch the issue where a teacher will be sent multiple emails in
one day.

A lot of this is generated by LLM tools (Claude), with some corrections
made manually.

* Fix whitespace and remove unneeded code/comments

* Mock out the config module for testing

* Update test dependencies

* Also include pytest-mock

* Add test case for multiple teachers with the same class