Add mock users, basic test for updating Alta groups (#37)
* Add test infrastructure for openPathUpdateAll

This adds two main pieces of code - the MockNeonUserBuilder and the
MockAltaUserBuilder. The Builder pattern allows us to create complex
objects (like fake user accounts) in a relatively readable way, and we
can then use those mock accounts to test out the behavior of real
functions without needing to connect to either Alta or Neon.

This also adds a test for the openPathUpdateAll method, but I'll be
working on cleaning that up to make it more readable.

* Added helper methods to make tests (slightly) more readable

* More refactoring, standardize variable names

* Add comments to explain the test

* Don't hardcode Alta IDs

* Avoid hardcoding OpenPathID in Neon users

* Also install pytz in the Github action (there's probably a better way to do this)

* Install requests in github action

* Add tests for dailyClassReminder.py (#39)

* Remove top-level API call to allow for mock responses

Top-level code is executed as soon as a module is imported, meaning that
it can't be mocked using our test framework. In order to write tests for
this code, we need to encapsulate the API call inside of a function so
that we can mock the API call rather than having it reach out to the
real API.

* Add passing tests for dailyClassReminders

This PR adds a MockEventBuilder and a few tests for the
dailyClassReminder script. All of these tests currently pass, so they
don't catch the issue where a teacher will be sent multiple emails in
one day.

A lot of this is generated by LLM tools (Claude), with some corrections
made manually.

* Fix whitespace and remove unneeded code/comments

* Mock out the config module for testing

* Update test dependencies

* Also include pytest-mock

* Add test case for multiple teachers with the same class

* Also install pytz in the Github action (there's probably a better way to do this)

* Remove Self, will re-add once I update the version of Python the tests are using